+++ src/control_plane/migration/redis/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(test)]
+mod tests;
+
+mod backup_redis_prefix;
+mod redis_backup_options;
+
+pub(crate) use backup_redis_prefix::backup_redis_prefix;
+pub(crate) use redis_backup_options::RedisBackupOptions;